Move PropertiesView into properties-page folder
